MS SQL driver rewritten to use ODBC interface via SQL Server 2008 Native Client; other minor fixes

+++ src/server/libnxsrv/snmp.cpp
@@ -97,7 +97,7 @@ DWORD LIBNXSRV_EXPORTABLE SnmpGet(DWORD dwVersion, SNMP_Transport *pTransport,
                {
 						int rawLen = (dwBufferSize - 1) / 2;
 						BYTE *raw = (BYTE *)malloc(rawLen);
-						rawLen = pVar->getRawValue(raw, rawLen);
+						rawLen = (int)pVar->getRawValue(raw, rawLen);
 						BinToStr(raw, rawLen, (TCHAR *)pValue);
 						free(raw);
                }
