+++ src/server/smsdrv/generic/pdumode.cpp
@@ -29,7 +29,7 @@ static bool SMSPack7BitChars(const char* input, char* output, int* outputLength,
 	int i;
 	unsigned char octet;
 	int used = 0;
-	const int inputLength = strlen(input);
+	const int inputLength = (int)strlen(input);
 
 	for (i = 0; i < inputLength; i++)
 	{
@@ -64,7 +64,7 @@ bool SMSCreatePDUString(const char* phoneNumber, const char* message, char* pduB
 	char payload[bufferSize];
 	char payloadHex[bufferSize*2 + 1];
 	int payloadSize = 0;
-	int phoneLength = strlen(phoneNumber);
+	int phoneLength = (int)strlen(phoneNumber);
 	int numberFormat = 0x91; // International format
 	int i;
